Display link to trend page, and have it use custom date ranges
The stats window in the list page will now show a link to the trends
page, which will also be able to parse custom date ranges.
when a custom date range translates into N days, it will
also display the stats for the previous N days, otherwise just
stats for the range picked.