Depend on admin oauth settings for private list access
Only OAuth portals that have been registered as authoritative
should be able to grant access to privatelists. This requires
a rewrite of all scripts that check access. Admin access has
also been cached, as there should be no reason to look that
up if it's already been acked by the portal.